backend/requirements.txt
fastapi==0.112.0
uvicorn[standard]==0.30.3
pydantic==2.8.2
python-multipart==0.0.9
# Upgraded OpenAI SDK for GPT-5 support
openai>=1.45.0
# Pin httpx to compatible version for OpenAI SDK
httpx>=0.27.2
openpyxl==3.1.5
rapidfuzz==3.9.7
chardet==5.2.0
# Additional dependencies for enhanced AI capabilities
python-dotenv>=1.0.0
# Async HTTP client for parallel processing
aiohttp>=3.9.1
# Data processing
pandas>=2.1.4
# Visualization libraries
matplotlib>=3.7.2
seaborn>=0.12.2
plotly>=5.17.0
# Machine learning
scikit-learn>=1.3.2
